renamed PackedItems to Deprecated for upcoming remodel
